# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FairChat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/check-username:
    post:
      operationId: checkUsername
      tags: [auth]
      summary: Check if username is available
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckUsernameRequest"
      responses:
        "200":
          description: Username availability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckUsernameResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Registration error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CheckUsernameRequest:
      type: object
      properties:
        username:
          type: string
      required:
        - username
    CheckUsernameResponse:
      type: object
      properties:
        available:
          type: boolean
      required:
        - available
    RegisterRequest:
      type: object
      properties:
        inviteCode:
          type: string
        displayName:
          type: string
        username:
          type: string
        password:
          type: string
        publicKey:
          type: string
      required:
        - inviteCode
        - displayName
        - username
        - password
        - publicKey
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        publicKey:
          type: string
        createdAt:
          type: string
          format: date-time
        isOnline:
          type: boolean
      required:
        - id
        - username
        - displayName
        - publicKey
        - createdAt
        - isOnline
...