# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FairChat API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Validate an invite code
 */
export const ValidateInviteCodeBody = zod.object({
  code: zod.string(),
});

export const ValidateInviteCodeResponse = zod.object({
  valid: zod.boolean(),
  invitedBy: zod.string().nullish(),
  invitedByDisplayName: zod.string().nullish(),
});

/**
 * @summary Check if username is available
 */
export const CheckUsernameBody = zod.object({
  username: zod.string(),
});

export const CheckUsernameResponse = zod.object({
  available: zod.boolean(),
});

/**
 * @summary Register a new user
 */
export const RegisterBody = zod.object({
  inviteCode: zod.string(),
  displayName: zod.string(),
  username: zod.string(),
  password: zod.string(),
  publicKey: zod.string(),
});

/**
 * @summary Login with username and password
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    username: zod.string(),
    displayName: zod.string(),
    publicKey: zod.string(),
    createdAt: zod.date(),
    isOnline: zod.boolean(),
  }),
});

/**
 * @summary Logout current user
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get current authenticated user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  displayName: zod.string(),
  publicKey: zod.string(),
  createdAt: zod.date(),
  isOnline: zod.boolean(),
});

/**
 * @summary Search users by username
 */
export const SearchUsersQueryParams = zod.object({
  q: zod.coerce.string(),
});

export const SearchUsersResponseItem = zod.object({
  id: zod.number(),
  username: zod.string(),
  displayName: zod.string(),
  publicKey: zod.string(),
  createdAt: zod.date(),
  isOnline: zod.boolean(),
});
export const SearchUsersResponse = zod.array(SearchUsersResponseItem);

/**
 * @summary Get a user by username
 */
export const GetUserByUsernameParams = zod.object({
  username: zod.coerce.string(),
});

export const GetUserByUsernameResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  displayName: zod.string(),
  publicKey: zod.string(),
  createdAt: zod.date(),
  isOnline: zod.boolean(),
});

/**
 * @summary Get all conversations for current user
 */
export const GetConversationsResponseItem = zod.object({
  user: zod.object({
    id: zod.number(),
    username: zod.string(),
    displayName: zod.string(),
    publicKey: zod.string(),
    createdAt: zod.date(),
    isOnline: zod.boolean(),
  }),
  lastMessage: zod
    .object({
      id: zod.number(),
      senderId: zod.number(),
      recipientId: zod.number(),
      encryptedContent: zod.string(),
      createdAt: zod.date(),
    })
    .nullish(),
  unreadCount: zod.number(),
});
export const GetConversationsResponse = zod.array(GetConversationsResponseItem);

/**
 * @summary Get messages with a specific user
 */
export const GetMessagesParams = zod.object({
  recipientId: zod.coerce.number(),
});

export const getMessagesQueryLimitDefault = 50;

export const GetMessagesQueryParams = zod.object({
  limit: zod.coerce.number().default(getMessagesQueryLimitDefault),
  before: zod.coerce.number().optional(),
});

export const GetMessagesResponseItem = zod.object({
  id: zod.number(),
  senderId: zod.number(),
  recipientId: zod.number(),
  encryptedContent: zod.string(),
  createdAt: zod.date(),
});
export const GetMessagesResponse = zod.array(GetMessagesResponseItem);

/**
 * @summary Send a message to a user
 */
export const SendMessageParams = zod.object({
  recipientId: zod.coerce.number(),
});

export const SendMessageBody = zod.object({
  encryptedContent: zod.string(),
});

/**
 * @summary List invite codes created by current user
 */
export const ListMyInviteCodesResponseItem = zod.object({
  id: zod.number(),
  code: zod.string(),
  createdBy: zod.number(),
  usedBy: zod.number().nullish(),
  usedAt: zod.date().nullish(),
  createdAt: zod.date(),
});
export const ListMyInviteCodesResponse = zod.array(
  ListMyInviteCodesResponseItem,
);